Fixed Project 2 catalog.sqlite, Finish project 3
Also ran some tests on project 1's catalog

+++ Project-3/code/DBFile.h
@@ -15,7 +15,7 @@ class DBFile {
 private:
 	File file;
 	string fileName;
-	int currPage;
+	off_t currPage;
 	Page page;
 
 public:
